Separate @maven dependencies from Copybara by allowing custom repository mapping (#1361)
* Separate @maven dependencies from Copybara by allowing custom repository mapping

The rationale is following:
When trying to upgrade the Maven artifacts, the current setup resulted in a conflict of `testonly` settings.
More precisely, a lib listed in `COPYBARA_MAVEN_ARTIFACTS` as `testonly` may be required by some non-`testonly` lib,
causing Bazel to refuse the build.
With the separation, these two sets of dependencies won't interfere with each other.
As a small downside, some common libs will be downloaded twice: for Copybara, and for the rest of the repo.

Alternative solutions considered:
* Patch the `COPYBARA_MAVEN_ARTIFACTS` list in the `WORKSPACE` file before feeding it to `maven_install`. This likely requires extra function, and list of artifacts to overwrite.
* Patch the `repositories.maven.bzl` of the Copybara repo, e.g. removing `testonly` attributes, or adjusting versions as needed. In this case we'd have three places in which the Maven deps are configured (Copybara repo, patches, WORKSPACE file).
* Provide the dependencies twice (imported from `COPYBARA_MAVEN_ARTIFACTS`, and then a copy of the same dependency with proper setup), hoping that some undocumented quirk of `maven_install` would resolve it as we wish.

* format

---------

Co-authored-by: aszady <aszady@users.noreply.github.com>
GitOrigin-RevId: 022befc0ae51198cebf16f52cbcd8697e0a595bf